add docstring to exported functions and structs; cleanup; update file names
