+++ bitcoinspv/helpers.go
