+++ bitcoinspv/types/block_event.go
@@ -14,12 +14,14 @@ const (
 	BlockConnected
 )
 
+// BlockEvent represents a new block event from subscription
 type BlockEvent struct {
 	EventType EventType
 	Height    int64
 	Header    *wire.BlockHeader
 }
 
+// NewBlockEvent creates and returns a new BlockEvent
 func NewBlockEvent(eventType EventType, height int64, header *wire.BlockHeader) *BlockEvent {
 	return &BlockEvent{
 		EventType: eventType,
